day5/secure-document-agent/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
google-generativeai==0.3.2
python-magic==0.4.27
cryptography>=42.0.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
pillow==10.1.0
pypdf==3.17.0
python-docx==1.1.0
openpyxl==3.1.2
pydantic==2.5.0
pydantic-settings>=2.0.0
structlog==23.2.0
